7-8-9/requirements.txt
flask
psycopg2-binary
requests
graphviz